refactor : util name
